requirements.txt
pathway
streamlit
python-dotenv
codecarbon
sentence-transformers
requests
numpy
